@@ -1,73 +1,58 @@
-import React, { Component } from 'react';
-import axios from 'axios';
-import { Consumer } from '../../context';
+import React, { useState, useEffect, useContext } from "react";
+import axios from "axios";
+import { Context } from "../../context";
 
-class Search extends Component {
-  state = {
-    trackTitle: ''
-  };
+const Search = () => {
+  const [state, setState] = useContext(Context);
+  const [userInput, setUserInput] = useState("");
+  const [trackTitle, setTrackTitle] = useState("");
 
-  findTrack = (dispatch, e) => {
-    e.preventDefault();
-
+  useEffect(() => {
     axios
       .get(
-        `https://cors-anywhere.herokuapp.com/http://api.musixmatch.com/ws/1.1/track.search?q_track=${
-          this.state.trackTitle
-        }&page_size=10&page=1&s_track_rating=desc&apikey=${
+        `https://cors-anywhere.herokuapp.com/http://api.musixmatch.com/ws/1.1/track.search?q_track=${trackTitle}&page_size=10&page=1&s_track_rating=desc&apikey=${
           process.env.REACT_APP_MM_KEY
         }`
       )
       .then(res => {
-        dispatch({
-          type: 'SEARCH_TRACKS',
-          payload: res.data.message.body.track_list
-        });
-
-        this.setState({ trackTitle: '' });
+        let track_list = res.data.message.body.track_list;
+        setState({ track_list: track_list, heading: "Search Results" });
       })
       .catch(err => console.log(err));
+  }, [trackTitle]);
+
+  const findTrack = e => {
+    e.preventDefault();
+    setTrackTitle(userInput);
   };
 
-  onChange = e => {
-    this.setState({ [e.target.name]: e.target.value });
+  const onChange = e => {
+    setUserInput(e.target.value);
   };
 
-  render() {
-    return (
-      <Consumer>
-        {value => {
-          const { dispatch } = value;
-          return (
-            <div className="card card-body mb-4 p-4">
-              <h1 className="display-4 text-center">
-                <i className="fas fa-music" /> Search For A Song
-              </h1>
-              <p className="lead text-center">Get the lyrics for any song</p>
-              <form onSubmit={this.findTrack.bind(this, dispatch)}>
-                <div className="form-group">
-                  <input
-                    type="text"
-                    className="form-control form-control-lg"
-                    placeholder="Song title..."
-                    name="trackTitle"
-                    value={this.state.trackTitle}
-                    onChange={this.onChange}
-                  />
-                </div>
-                <button
-                  className="btn btn-primary btn-lg btn-block mb-5"
-                  type="submit"
-                >
-                  Get Track Lyrics
-                </button>
-              </form>
-            </div>
-          );
-        }}
-      </Consumer>
-    );
-  }
-}
+  return (
+    <div className="card card-body mb-4 p-4">
+      <h1 className="display-4 text-center">
+        <i className="fas fa-music" /> Search For A Song
+      </h1>
+      <p className="lead text-center">Get the lyrics for any song</p>
+      <form onSubmit={findTrack}>
+        <div className="form-group">
+          <input
+            type="text"
+            className="form-control form-control-lg"
+            placeholder="Song title..."
+            name="userInput"
+            value={userInput}
+            onChange={onChange}
+          />
+        </div>
+        <button className="btn btn-primary btn-lg btn-block mb-5" type="submit">
+          Get Track Lyrics
+        </button>
+      </form>
+    </div>
+  );
+};
 
 export default Search;
