@@ -1,6 +1,6 @@
-import React from 'react';
-import Tracks from '../tracks/Tracks';
-import Search from '../tracks/Search';
+import React from "react";
+import Tracks from "../tracks/Tracks";
+import Search from "../tracks/Search";
 
 const Index = () => {
   return (
