@@ -1,33 +1,26 @@
-import React, { Component } from 'react';
-import { Consumer } from '../../context';
-import Spinner from '../layout/Spinner';
-import Track from '../tracks/Track';
+import React, { useContext } from "react";
+import { Context } from "../../context";
+import Spinner from "../layout/Spinner";
+import Track from "../tracks/Track";
 
-class Tracks extends Component {
-  render() {
+const Tracks = () => {
+  const [state] = useContext(Context);
+  const { track_list, heading } = state;
+
+  if (track_list === undefined || track_list.length === 0) {
+    return <Spinner />;
+  } else {
     return (
-      <Consumer>
-        {value => {
-          const { track_list, heading } = value;
-
-          if (track_list === undefined || track_list.length === 0) {
-            return <Spinner />;
-          } else {
-            return (
-              <React.Fragment>
-                <h3 className="text-center mb-4">{heading}</h3>
-                <div className="row">
-                  {track_list.map(item => (
-                    <Track key={item.track.track_id} track={item.track} />
-                  ))}
-                </div>
-              </React.Fragment>
-            );
-          }
-        }}
-      </Consumer>
+      <>
+        <h3 className="text-center mb-4">{heading}</h3>
+        <div className="row">
+          {track_list.map(item => (
+            <Track key={item.track.track_id} track={item.track} />
+          ))}
+        </div>
+      </>
     );
   }
-}
+};
 
 export default Tracks;
