@@ -1,29 +1,31 @@
-import React, { Component } from 'react';
-import axios from 'axios';
+import React, { useState, useEffect } from "react";
+import axios from "axios";
 
-const Context = React.createContext();
+export const Context = React.createContext();
 
-const reducer = (state, action) => {
-  switch (action.type) {
-    case 'SEARCH_TRACKS':
-      return {
-        ...state,
-        track_list: action.payload,
-        heading: 'Search Results'
-      };
-    default:
-      return state;
-  }
-};
+// const reducer = (state, action) => {
+//   switch (action.type) {
+//     case 'SEARCH_TRACKS':
+//       return {
+//         ...state,
+//         track_list: action.payload,
+//         heading: 'Search Results'
+//       };
+//     default:
+//       return state;
+//   }
+// };
 
-export class Provider extends Component {
-  state = {
+export function ContextController({ children }) {
+  let intialState = {
     track_list: [],
-    heading: 'Top 10 Tracks',
-    dispatch: action => this.setState(state => reducer(state, action))
+    heading: ""
+    // dispatch: action => this.setState(state => reducer(state, action))
   };
 
-  componentDidMount() {
+  const [state, setState] = useState(intialState);
+
+  useEffect(() => {
     axios
       .get(
         `https://cors-anywhere.herokuapp.com/http://api.musixmatch.com/ws/1.1/chart.tracks.get?page=1&page_size=10&country=us&f_has_lyrics=1&apikey=${
@@ -32,18 +34,15 @@ export class Provider extends Component {
       )
       .then(res => {
         // console.log(res.data);
-        this.setState({ track_list: res.data.message.body.track_list });
+        setState({
+          track_list: res.data.message.body.track_list,
+          heading: "Top 10 Tracks"
+        });
       })
       .catch(err => console.log(err));
-  }
+  }, []);
 
-  render() {
-    return (
-      <Context.Provider value={this.state}>
-        {this.props.children}
-      </Context.Provider>
-    );
-  }
+  return (
+    <Context.Provider value={[state, setState]}>{children}</Context.Provider>
+  );
 }
-
-export const Consumer = Context.Consumer;
